requirements.txt
streamlit
matplotlib
fpdf2
python-dotenv
schedule
requests
